requirements.txt
fastapi
uvicorn
python-socketio
websockets
requests
pydub
googletrans
SpeechRecognition
murf
streamlit
sounddevice
numpy
websocket-client
pydub
python-mulitpart
authlib
python-dotenv
imageio-ffmpeg